% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_wrappers.R
\name{get_all_manifests}
\alias{get_all_manifests}
\title{Call `storage/project/manifests` Schematic endpoint for a given list of storage projects (output from `storage/projects`) and output information as a dataframe}
\usage{
get_all_manifests(
  storage_project_list,
  asset_view,
  input_token,
  base_url = "https://schematic.dnt-dev.sagebase.org",
  verbose = FALSE
)
}
\arguments{
\item{storage_project_list}{List output from `storage_projects` schematic endpoint}

\item{asset_view}{ID of view listing all project data assets. For example, for Synapse this would be the Synapse ID of the fileview listing all data assets for a given project.(i.e. master_fileview in config.yml)}

\item{input_token}{Synapse PAT}

\item{base_url}{Base URL of schematic API}

\item{verbose}{T/F for console messages}
}
\description{
Call `storage/project/manifests` Schematic endpoint for a given list of storage projects (output from `storage/projects`) and output information as a dataframe
}
